[core] reduce test iteration to fix time degradation (#33039)
[core] reduce test iteration to fix time degradation

This commit is to enable CI tests to run in a resonable time on Windows
debug build.

There's a problem with function ov::CoreImpl::get_plugin which is
spotted on high count of iteration in multithread run, the locking
mechanisms used waste big time on read and write.

### Details:
- **CoreThreadingTests.RegisterPlugin** test can spanning arbitrary
number of threads, by default are 8 threads. the bottleneck seems to be
the extensive use of locks in terms of construction and destruction of
objects which cause the test to timeout when using high number of
iteration per thread.

### Tickets:
 - CVS-176899

Signed-off-by: Walid Aly <walid.aly@intel.com>