[TRANSFORMATIONS] Fix incorrect conditions of conversion for ov::Range in cases when FP16 compression is enabled (#33051)
### Details:
Originally, compression to lower precisions was disabled for ov::Range
in dcf9605 to prevent accuracy loss for the llama model. However, the
commit aimed at the case of compression to FP16 only, but mistakenly
disabled compression of integral types as well.

Fix the check in the dedicated function of ov::Range to prevent the
conversion to FP16 only.

### Tickets:
 - [CVS-176258](https://jira.devtools.intel.com/browse/CVS-176258)

Signed-off-by: Andrii Staikov <andrii.staikov@intel.com>