[GPU] Fix roll wrong axis, shift handling issue. (#33091)
### Description of the issue(symptom, root-cause, how it was resolved)
 Unexpected behavior at negative axis in roll primitive with 3D shape.
From
[roll-7](https://docs.openvino.ai/2024/documentation/openvino-ir-format/operation-sets/operation-specs/movement/roll-7.html),
axis and shift support negative value but current roll has the problem
for handling 3D shape.

 - How it was resolved
Normalized the axis during roll primitive creation.
Updating the axes in get_kernel_params() complicates the problem because
canoicalize_shapes changs the shape.




#### Reproduction step and snapshot (if applicable. Do not attach for
customer model)
$ python lfm2.py   // Basic script is in the ticket.
...
<|startoftext|><|im_start|>user
What is C. elegans?<|im_end|>
<|im_start|>assistant
C. C. C. The nematode a. the nematodes is a. a. a. a. a. a. a. a. a. a.
a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a. a.
a. a. a. a. a. a. a. a.

#### Problematic graph 
<img width="766" height="955" alt="image"
src="https://github.com/user-attachments/assets/6b4185fb-59a3-4058-99b2-c7435f42c1c8"
/>


#### Checklist 
 - [x] Is it a proper fix?
 - [x] Did you include test case for this fix, if necessary? 
- [x] Did you review existing test that can be extended to cover this
scenario? Which test did you review?

### Tickets:
 - *CVS-176974*

---------

Signed-off-by: hyunback <hyunback.kim@intel.com>