[Transformation] Identity with Result removal transformation workaround (#33175)
### Details:
- Now correctly removes all Identity op from graph even in Result edge
cases where the input node has another Result attached to it
- Unlocks https://github.com/openvinotoolkit/openvino/pull/33166

### Tickets:
 - 177222

### Explaination:
When trying to convert graph:
```
p0 -> I0 -----\
                Multiply ---- Result1
p1 -> I1 -----/           \
                           I2 ----- Result2
```
to
```
p0 ----\
        Multiply ---- Result1
p1 ----/      \
               Result2
```
the issue arises since Multiply outputs 1 tensor: T1, with a name "t1".
Since Identity is a copy operation, then the I2 outputs a new tensor T2
with a name "t2". After removing I2, we need to keep T2, but Multiply
has only one output, T1. The possible solutions were:

- Keep only Multiply’s name "t1": Result2 now would have to have source
tensor renamed to "t1"
- Keep only Identity’s name "t2": Result1 Result2 now would have to have
source tensor renamed to "t2"
- Merge (union) them: both old names still exist, but the tensor’s name
set changed vs the original one
- And the fourth one, replace I2 with another OP. This would generate a
new tensor "t2" from "t1" and if the new OP effectively "does nothing",
the names would be kept

So for now, there is no satisfactory solution. Therefore a workaround
for the NPU plugin is to use another OP: in this PR I decided to replace
I2 edge case with Reshape to have the marginal performance loss, if any.
It won't be eliminated with optimizations for the same reasons why I2
cannot be eliminated.

---------

Co-authored-by: Mikhail Ryzhov <mikhail.ryzhov@intel.com>