Correct use Tensor::data() for model inputs when read-only tensors (#33120)
### Details:
- The tensor can be created as read-only view when input pointer is
constant. This kind of tensor can be used as model inputs as model input
should be read-only. There is issue to use `data()` if input tensor is
not const object the acquired pointer is not constant which throw
exception to indicate that there is intention to modify of read-only
tensor.
- Update core, and plugins to convert tensor object to const when get
data pointer to call correct overload
- Update tests to use read-only tensor for inference on different
devices.
 
 ### Blocked PR:
 - openvinotoolkit/openvino.genai#3086

### Tickets:
 - CVS-176653

---------

Signed-off-by: Pawel Raasz <pawel.raasz@intel.com>
Signed-off-by: Raasz, Pawel <pawel.raasz@intel.com>