[TensorFlow Layer Test] TimeoutError OpenVINO running timeout (#33225)
The test which was timing out used a seemingly constant input data seed.
It was global though, so when PyTest started running testcases
concurrently, the generated data in a given testcase depended on the
order in which the testcase run (and this was random).

The fix makes the stimuli generator local, so that each test case has a
guaranteed initial state.

Ticket: CVS-176864