[NPU][IE-MDK] Add fallbacks for platform indentification (#33156)
### Details:
- The skip config gets multiple fallbacks for platform identification.
- Some tests can set their own desired platform, while the skip config
is populated before any test is run. This means that it needs to know
what's the expected platform, in order to populate the list correctly.
- This implementation uses `IE_NPU_TESTS_PLATFORM` as the first source
form platform identification, as it's a mandatory envvar to run
`ov_npu_func_tests`. It then falls back to using
`IE_NPU_TESTS_DEVICE_NAME`. If neither envvar provide a usable platform,
the final fallback is the list returned by the NPU Plugin

### Tickets:
 - E 189860