[NPU] Integrate UD48 compiler in NPU plugin (#33129)
### Details:
- Add cmake option "ENABLE_INTEL_NPU_COMPILER" to enable/disable the
download of the library. Default value is ON.
- Library is downloaded from the openvino storage, but the original
source is [NPU Compiler
Release](https://github.com/openvinotoolkit/npu_compiler/releases).
- While the compiler will be included by default in openvino packages,
it will not be used yet by default by the NPU plugin.
- The new compiler library inside plugin can only be used through
NPU_COMPILER_TYPE=PLUGIN.
- New tests will gradually be migrated to use compiler-in-plugin flow (
outside of this PR since part of the scripts are outside of this
repository).
- Logic in NPU Plugin to automatically select the suitable compiler type
will be added later as well.

### Tickets:
 - *CVS-174281*

---------

Signed-off-by: Kang, Wenjing <wenjing.kang@intel.com>
Co-authored-by: Kang, Wenjing <wenjing.kang@intel.com>