[NPU] UMD caching: pass the model hash to the driver (#33046)
### Details:
- The NPU plugin sends its own model hash to the driver if the driver
supports this feature. This hash is computed using the pass provided by
OV. The driver will use this hash as part of the computation for the
final hash. The final hash determines a UMD cache hit/miss.
- Reorganized the pipeline that runs for serializing the model to assure
correctness. First, all operator conversion passes are run. Then runtime
information (such as the `WeightlessCacheAttribute`) is stored. Finally,
the model is serialized and the hash is computed if the feature is
supported.

### Tickets:
 - *EISW-178935*